Final Version

+++ pages/3_Couches.py
@@ -26,7 +26,7 @@ def Type1():
         df = df.reset_index()
         df = df[['Source','Maximum Price','Minimum Price','Average','# of Products']]
         col1, col2, col3 = st.columns(3)
-        col1.metric("Average Market Price", df['Average'].mean())
+        col1.metric("Average Market Price", df['Average'].mean().astype('int'))
         col2.metric("Units on Market", df['# of Products'].sum())
         col3.metric("Time on Market", "10 Days")
         col1 = st.columns(1)
@@ -35,7 +35,7 @@ def Type1():
         df = load_data('RepairCosts')
         df = df[df.Product == 'Couch']
         col1, col2, col3 = st.columns(3)
-        col1.metric("Average Repair Cost", df['Average Cost'].mean())
+        col1.metric("Average Repair Cost", df['Average Cost'].mean().astype('int'))
         col2.metric("Potential Issues", df.count().mean().astype('str'))
         col3.metric("Estimated Labor", "10 Days")
         col1 = st.columns(1)
@@ -74,7 +74,7 @@ def Type2():
         df = df.reset_index()
         df = df[['Source','Maximum Price','Minimum Price','Average','# of Products']]
         col1, col2, col3 = st.columns(3)
-        col1.metric("Average Market Price", df['Average'].mean())
+        col1.metric("Average Market Price", df['Average'].mean().astype('int'))
         col2.metric("Units on Market", df['# of Products'].sum())
         col3.metric("Time on Market", "10 Days")
         col1 = st.columns(1)
@@ -83,7 +83,7 @@ def Type2():
         df = load_data('RepairCosts')
         df = df[df.Product == 'Couch']
         col1, col2, col3 = st.columns(3)
-        col1.metric("Average Repair Cost", df['Average Cost'].mean())
+        col1.metric("Average Repair Cost", df['Average Cost'].mean().astype('int'))
         col2.metric("Potential Issues", df.count().mean().astype('str'))
         col3.metric("Estimated Labor", "10 Days")
         col1 = st.columns(1)
@@ -120,7 +120,7 @@ def Type3():
         df = df.reset_index()
         df = df[['Source','Maximum Price','Minimum Price','Average','# of Products']]
         col1, col2, col3 = st.columns(3)
-        col1.metric("Average Market Price", df['Average'].mean())
+        col1.metric("Average Market Price", df['Average'].mean().astype('int'))
         col2.metric("Units on Market", df['# of Products'].sum())
         col3.metric("Time on Market", "10 Days")
         col1 = st.columns(1)
